+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -3,16 +3,17 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.iteractionapi.dto.OrderDto;
 import ru.yandex.practicum.iteractionapi.dto.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
     PaymentDto payment(OrderDto orderDto);
 
-    Double getTotalCost(OrderDto orderDto);
+    BigDecimal getTotalCost(OrderDto orderDto);
 
     void paymentSuccess(UUID uuid);
 
-    Double productCost(OrderDto orderDto);
+    BigDecimal productCost(OrderDto orderDto);
 
     void paymentFailed(UUID uuid);
 }
